Creating footer cookie banner inhtml and changing it dynamically in js
Using JavaScript and DOM manipulation, simulate the ability for a user to "accept cookie usage". Provide a footer banner where a user can click a link to "Accept Cookies." When the user clicks "Accept Cookies", change the content on the banner to display new content in the footer stating: "Cookies were accepted. Would you like to revoke?" and provide a link for the user to revoke. If the user chooses to revoke, use JavaScript to revert to the original banner content.